fix(pagination): demo拆解与规范
